+++ main.tf
@@ -224,6 +224,41 @@ resource "azurerm_container_app" "container_app" {
         }
       }
     }
+    dynamic "custom_scale_rule" {
+      for_each = each.value.template.custom_scale_rule == null ? [] : each.value.template.custom_scale_rule
+
+      content {
+        custom_rule_type = custom_scale_rule.value.custom_rule_type
+        metadata         = custom_scale_rule.value.metadata
+        name             = custom_scale_rule.value.name
+
+        dynamic "authentication" {
+          for_each = custom_scale_rule.value.authentication == null ? [] : custom_scale_rule.value.authentication
+
+          content {
+            secret_name       = authentication.value.secret_name
+            trigger_parameter = authentication.value.trigger_parameter
+          }
+        }
+      }
+    }
+    dynamic "http_scale_rule" {
+      for_each = each.value.template.http_scale_rule == null ? [] : each.value.template.http_scale_rule
+
+      content {
+        concurrent_requests = http_scale_rule.value.concurrent_requests
+        name                = http_scale_rule.value.name
+
+        dynamic "authentication" {
+          for_each = http_scale_rule.value.authentication == null ? [] : http_scale_rule.value.authentication
+
+          content {
+            secret_name       = authentication.value.secret_name
+            trigger_parameter = authentication.value.trigger_parameter
+          }
+        }
+      }
+    }
     dynamic "init_container" {
       for_each = each.value.template.init_containers == null ? [] : each.value.template.init_containers
 
@@ -261,38 +296,6 @@ resource "azurerm_container_app" "container_app" {
         name         = volume.value.name
         storage_name = volume.value.storage_name
         storage_type = volume.value.storage_type
-      }
-    }
-    dynamic "http_scale_rule" {
-      for_each = each.value.template.http_scale_rule == null ? [] : each.value.template.http_scale_rule
-      content {
-        name                = http_scale_rule.value.name
-        concurrent_requests = http_scale_rule.value.concurrent_requests
-        dynamic "authentication" {
-          for_each = http_scale_rule.value.authentication == null ? [] : http_scale_rule.value.authentication
-          content {
-            secret_name       = authentication.value.secret_name
-            trigger_parameter = authentication.value.trigger_parameter
-          }
-        }
-      }
-    }
-
-    dynamic "custom_scale_rule" {
-      for_each = each.value.template.custom_scale_rule == null ? [] : each.value.template.custom_scale_rule
-      content {
-        custom_rule_type = custom_scale_rule.value.custom_rule_type
-        metadata         = custom_scale_rule.value.metadata
-        name             = custom_scale_rule.value.name
-
-        dynamic "authentication" {
-          for_each = custom_scale_rule.value.authentication == null ? [] : custom_scale_rule.value.authentication
-
-          content {
-            secret_name       = authentication.value.secret_name
-            trigger_parameter = authentication.value.trigger_parameter
-          }
-        }
       }
     }
   }
@@ -362,5 +365,4 @@ resource "azurerm_container_app" "container_app" {
       value = local.container_app_secrets[each.key][secret.key]
     }
   }
-
 }